Lab 6: Copy and update the register macro
This step creates the registers and enables the register
visualization in makerchip's VIZ view.

The macro creates 32 registers of 32-bits each (since we're
implementing the 32-bit subset of the RISC-V ISA), wiring
the registers with 1 write channel and 2 read channels.

Each of those channels has an enable pin, which is mapped
to the appropriate `$r*_valid` value from earlier:
`$rd` is the register destination (where an instruction will
write to), with `$rd_valid` controlling whether the destination
will actually be written to;
`$rs1` is source register one, the register number of the first
input to the instruction, with `$rs1_valid` similarly controlling
whether the instruction actually has a first input value;
`$rs2` is source register two, the register number of the second
input to the instruction, with `$rs2_valid` similarly controlling
whether the instruction actually has a second input value.

This lab expects the actual bits that are mapped to `rs1` and `rs2`
to be `$src1_value` and `$src2_value`, so we're sticking with that,
though `rd_data` was my own choice for consistency.