Lab 7: Implement the ALU
This is an extremely minimal ALU, implementing only ADD
and ADDI, and isn't yet completely functional: the result
is stored locally but not written back to the register file.

`$result` is always set, even if this is an instruction which
doesn't use the ALU.  In those cases, a literal `0` is captured
to the `result` instead, with the expectation it won't ever be
written out anyway.