Lab 8: Write the result back (unless it's to x0)
These two changes aren't strictly related, but they are in the same
lab.

`$rd_valid` controls whether data should be written to the register
identified by `$rd`.  The value of `$result` is written to that
register by assignment here.

Since `x0` should always contain zero, the simplest implementation
to prevent overwriting that zero is to ensure `$rd_valid` is always
false if the destination register is zero.