Lab 9: Handle branching and *passed
Branching is simply a matter of arithmetic that updates
`$next_pc`, so `$next_pc` has been moved to the bottom of the code
block.

It's now set to the appropriate value based on the type of branching.

Now that register reads, register writes, and branches all work
appropriately, the `*passed` value can be set appropriately.  The
`m4+tb()` macro handles that.